Initial commit
Minimal desired functionality: weekdays-aware habit tracking!

* proper tracking of .+ type habits
  * weekdays that are not in the HABIT_WEEKDAYS property are omitted
    in agenda habit display as if they haven't existed; effectively, a week shrinks to only those
    weekdays that are in the property.
  * a function for `after-todo-state-change-hook' to adjust rescheduling
    according to HABIT_WEEKDAYS
* proper tracking of ++ type habits
  * either I didn't understand them right or they were not working
    properly in the original org-habit module
  * nevertheless they are displayed exactly the way I wanted them to be
    displayed

Aside of a bunch of several new auxiliary functions, two original functions
`org-habit-parse-todo' and `org-habit-get-faces' were patched and
`org-habit-build-graph' was heavily rewritten to achieve the desired
behavior. As a result, they are no longer compatible with original
versions because `org-habit-parse-todo' produces slightly different
datatype whereas `org-habit-get-faces' expects it. The latter is called
from within `org-habit-build-graph' and thus this new datatype is being
used there internally as well, although the interface of this function
didn't change. The mentioned datatype is nothing particular, but alist
of (days . weekday) pairs, where days is the number of days
since epoch and weekday is number of day in a week (Monday - 1,
Sunday - 7). Originally simple lists of days were used to store and
process habit information.

* TODO
  * "+" type habits - what they are (for)
  * different glyphs for different done states (if several) on a habit
    graph, to distinguish between, for example, DONE and CANCELLED;
    maybe different colors as well.
  * minor user-friendly options, such as locale weekdays
    names (Mon, Tue..), etc.
  * code cleaning